# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ashva Experts API (Consolidated)
  version: 0.2.0
  description: |
    Consolidated API specification for manual testing during development.

    Notes:
    - Base URL for local testing via Nginx: `http://localhost:8080`
    - All `/api/v1/*` routes are served through the gateway.
    - Auth endpoints under `/api/v1/auth/*` are public.
    - Non-auth endpoints require `Authorization: Bearer <access_token>`.
servers:
  - url: http://localhost:8080
    description: Local via Nginx (recommended)
  - url: http://localhost:8000
    description: Local gateway direct
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CouponCreateRequest:
      type: object
      properties:
        kind: { type: string, enum: [promo], example: promo }
        code:
          type: string
          nullable: true
          description: If omitted, service auto-generates a code.
          example: diwali2026_ABC123XYZ9
        campaign_name: { type: string, nullable: true, example: diwali2026 }
        discount_type: { type: string, enum: [percent, fixed], example: percent }
        discount_value: { type: number, example: 10 }
        applies_to:
          type: array
          items: { type: string }
          nullable: true
          example: [rental, amc]
        min_amount: { type: number, nullable: true, example: 499 }
        max_redemptions: { type: integer, nullable: true, example: 1000 }
        per_user_limit: { type: integer, nullable: true, example: 1 }
        valid_from: { type: string, format: date-time, nullable: true }
        valid_to: { type: string, format: date-time, nullable: true }
        active: { type: boolean, example: true }
      required: [kind, discount_type, discount_value]
    CouponResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        kind: { type: string, enum: [promo, referral] }
        campaign_name: { type: string, nullable: true }
        discount_type: { type: string, enum: [percent, fixed] }
        discount_value: { type: number }
        applies_to:
          type: array
          items: { type: string }
          nullable: true
        min_amount: { type: number, nullable: true }
        max_redemptions: { type: integer, nullable: true }
        per_user_limit: { type: integer, nullable: true }
        redeemed_count: { type: integer }
        valid_from: { type: string, format: date-time, nullable: true }
        valid_to: { type: string, format: date-time, nullable: true }
        active: { type: boolean }
        referrer_user_id: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, code, kind, discount_type, discount_value, redeemed_count, active, created_at, updated_at]
    CouponListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/CouponResponse" }
        total: { type: integer }
      required: [items, total]
    ReferralProgramUpdateRequest:
      type: object
      properties:
        referrer_credit_amount:
          type: number
          description: Fixed credit amount (one-time) for the referrer, applied on next rental renewal.
          example: 100
        referred_percent: { type: integer, example: 5 }
        active: { type: boolean, example: true }
      required: [referrer_credit_amount, referred_percent]
    ReferralProgramResponse:
      type: object
      properties:
        referrer_credit_amount: { type: number }
        referred_percent: { type: integer }
        active: { type: boolean }
        created_at: { type: string, format: date-time }
      required: [referrer_credit_amount, referred_percent, active, created_at]
    ReferralGenerateResponse:
      type: object
      properties:
        code: { type: string, example: REFABC123XYZ9 }
      required: [code]
...
paths:
...
  /api/v1/coupons:
    get:
      tags: [coupons]
      summary: List coupons (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: kind
          schema: { type: string, enum: [promo, referral] }
        - in: query
          name: active
          schema: { type: boolean }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CouponListResponse" }
    post:
      tags: [coupons]
      summary: Create coupon (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CouponCreateRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CouponResponse" }
  /api/v1/coupons/referrals/generate:
    post:
      tags: [coupons]
      summary: Generate referral code (subscriber)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReferralGenerateResponse" }
  /api/v1/coupons/referrals/program:
    put:
      tags: [coupons]
      summary: Set referral program percentages (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReferralProgramUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReferralProgramResponse" }
...